add mongo

+++ src/blogs/routers/handlers/delete-blog.handler.ts
@@ -4,14 +4,24 @@ import { createErrorMessages } from '../../../core/utils/error.utils';
 import { blogsRepository } from "../../repositories/blog.repository";
 
 
-export function deleteBlogHandler(req: Request<{id: string}>, res: Response) {
-    const id = req.params.id;
-    const blog = blogsRepository.findById(id);
+export async function deleteBlogHandler(req: Request<{id: string}>, res: Response) {
+   try {
+       const id = req.params.id;
+       const blog = await blogsRepository.findById(id);
 
-    try {
-        blogsRepository.delete(id);
-        return res.sendStatus(HttpStatus.NoContent);
-    } catch (e) {
-        return res.sendStatus(HttpStatus.NotFound);
-    }
+       if (!blog) {
+           res
+               .status(HttpStatus.NotFound)
+               .send(
+                   createErrorMessages([{ field: 'id', message: 'Blog not found' }]),
+               );
+           return;
+       }
+
+       await blogsRepository.delete(id);
+       return res.sendStatus(HttpStatus.NoContent);
+   } catch (e) {
+           return res.sendStatus(HttpStatus.InternalServerError);
+       }
+
 }
